backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.12.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
Pillow==10.1.0
bcrypt==4.1.1
email-validator==2.1.0
